use mmc1
is contingent on a change in the nesfab codebase (relaxing default rom
sizes), which i'm not committing because i think pubby can merge that in
fine (and if not then this change isn't stable), so there'll be a bit of
a desync. If you want to replicate the change, go to `src/mapper.cpp`
and edit the mmc1 params.